split userlistelement in two
