+++ src/client/components/UserList/UserListElement/AdministrationListElement/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AdministrationListElement'
