+++ src/client/components/UserList/UserListElement/CollaboratorListElement/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CollaboratorListElement'
