--- src/client/components/UserList/UserListElement/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserListElement'
